Implemented the ability for a customer to know what lunches they have already bought for their children
SERVER
made it so that the account object sent by session/start contains a full featured purchases list

CUSTOMER
made it so that the main user display shows the list of days with lunches already purchased
added tooltip that lists all lunches in the database for the student
made it so that lunch selection tool disables buttons for days already purchased
disabled button shows a tooltip explaining why

SCHOOLADMIN
made it so that the system shows a count of the lunches already purchased for a student
created a tooltip for that count that lists all the lunches for the student
(the lunch selection tool is the same code as for customers so it already disables buttons)

switch from html5tooltips to qtip library. I really like html5tooltips but it's api is cumbersome and didn't work for my needs.